updated the DataManager Class to handle both locally installed dataset and openML
The DataManager class dynamically adapts weather a locally installed dataset or an OpenML dataset is used, depending on the parameters passed to the __init__ class. If dataset_name & target_col are provided, then a locally stored dataset (.csv / .pkl) will be loaded. If an dataset_id is provided, then OpenML will be queried for a dataset with this id. if all are provided: OpenML > local dataset